+++ R/kernels.R
@@ -3,9 +3,9 @@
 #'
 #' @return The value of the sinc function at each point in `u`.
 sinc <- function(u) {
-  ifelse(u == 0, 
+  ifelse(u == 0,
          1,  # When u = 0, limit of sin(u)/(pi*u) is 1
-         sin(u) / (pi * u))
+         sin(pi * u) / (pi * u))
 }
 
 #' Define the closed form FT of the infinite order kernel sin(x)/(pi*x)
@@ -15,7 +15,7 @@ sinc <- function(u) {
 #' @return The value of the Fourier Transform of the sinc function at each point in `x`.
 #'
 sinc_ft <- function(x) {
-  (sign(x + 1) - sign(x - 1)) / 2
+  (sign(x + pi) - sign(x - pi)) / 2
 }
 
 #' Define the Fourier transform of a infinite kernel proposed in Schennach 2004
@@ -70,7 +70,7 @@ W_kernel <- function(u, L = 10) {
 #' @return Returns the value of the Normal kernel function at the given input.
 #'
 normal_kernel <- function(u) {
-  exp(-u^2 / 2)
+  dnorm(u, mean = 0, sd = 1)
 }
 
 #' Fourier Transform of Normal Kernel
@@ -80,7 +80,7 @@ normal_kernel <- function(u) {
 #' @return Returns the value of the Fourier transform of the Normal kernel at the given frequency/frequencies.
 #'
 normal_kernel_ft <- function(xi) {
-  exp(-xi^2 / 2)
+  dnorm(xi, mean = 0, sd = 1) * (sqrt(2 * pi))
 }
 
 #' Epanechnikov Kernel
@@ -90,7 +90,7 @@ normal_kernel_ft <- function(xi) {
 #' @return Returns the value of the Epanechnikov kernel function at the given input.
 #'
 epanechnikov_kernel <- function(u) {
-  ifelse(abs(u) <= 1, 3 / 4 * (1 - u^2), 0)
+  ifelse(abs(u) <= 1, 0.75 * (1 - u^2), 0)
 }
 
 #' Fourier Transform Epanechnikov Kernel
@@ -101,4 +101,44 @@ epanechnikov_kernel <- function(u) {
 #'
 epanechnikov_kernel_ft <- function(xi) {
   (3 * sin(xi)) / (2 * xi) - (3 * (2 * xi * cos(xi) + (-2 + xi^2) * sin(xi))) / (2 * xi^3)
+}
+
+#' Create kernel functions based on configuration
+#'
+#' @param kernel.fun A string specifying the kernel function to be used.
+#' @param if_approx_kernel Logical. If TRUE, uses approximations for the kernel function.
+#' @param kernel.resol The resolution for kernel function approximation.
+#'
+#' @return A list containing kernel function, its Fourier transform, and the kernel type
+#'
+#' @export
+create_kernel_functions <- function(kernel.fun = "Schennach2004",
+                                   if_approx_kernel = TRUE,
+                                   kernel.resol = 1000
+                                   ) {
+
+  # Store the kernel type (same as input string or derived from function)
+  kernel_type <- kernel.fun
+
+  if (kernel.fun == "Schennach2004") {
+    if (if_approx_kernel) {
+      inf_k <- function(u) fun_approx(u, u_lb = -100, u_ub = 100, resol = kernel.resol)
+    } else {
+      inf_k <- W_kernel
+    }
+    inf_k_ft <- W_kernel_ft
+  } else if (kernel.fun == "sinc") {
+    inf_k <- sinc
+    inf_k_ft <- sinc_ft
+  } else if (kernel.fun == "normal") {
+    inf_k <- normal_kernel
+    inf_k_ft <- normal_kernel_ft
+  } else if (kernel.fun == "epanechnikov") {
+    inf_k <- epanechnikov_kernel
+    inf_k_ft <- epanechnikov_kernel_ft
+  } else {
+    stop("Unsupported kernel function: ", kernel.fun)
+  }
+
+  return(list(kernel = inf_k, kernel_ft = inf_k_ft, kernel_type = kernel_type))
 }