+++ R/rbbnp-pkg.R
@@ -1,7 +1,7 @@
 #' @import tidyr
 #' @import dplyr
 #' @import ggplot2
-#' @importFrom stats approxfun qnorm runif var rchisq rnorm
+#' @importFrom stats approxfun qnorm runif var rchisq rnorm IQR dnorm sd
 #' @importFrom gridExtra grid.arrange
 #' @export create_kernel_functions
 #' @export create_biasBound_config